chore: seperate functions
