+++ src/app.d.ts
@@ -1,13 +1,9 @@
 // See https://kit.svelte.dev/docs/types#app
-
-import type { LibSQLDatabase } from 'drizzle-orm/libsql';
-
 // for information about these interfaces
 declare global {
 	namespace App {
 		// interface Error {}
 		interface Locals {
-			db: LibSQLDatabase;
 			weekId: number;
 		}
 		interface PageData {
