+++ src/hooks.server.ts
@@ -1,13 +1,7 @@
-import { SECRET_DB_URL, SECRET_DB_AUTH_TOKEN } from '$env/static/private';
 import type { Handle } from '@sveltejs/kit';
-import { drizzle } from 'drizzle-orm/libsql';
-import { createClient } from '@libsql/client/http';
 import { getWeekId } from '$lib/server/weeks.service';
 
 export const handle: Handle = async ({ event, resolve }) => {
-	// load database
-	event.locals.db = drizzle(createClient({ url: SECRET_DB_URL, authToken: SECRET_DB_AUTH_TOKEN }));
-
 	// load weekId
 	event.locals.weekId = getWeekId()!;
